Refactor dtype propagation
ghstack-source-id: 38bff03b7d7122bd35722764f778b8176271c94e
Pull Request resolved: https://github.com/pytorch/pytorch/pull/139945

+++ torch/_inductor/codegen/triton.py
@@ -1255,7 +1255,7 @@ class TritonKernelOverrides(TritonOverrides):
             return cse_val
 
         mantissa = V.kernel.cse.newvar(dtype=x.dtype)
-        exponent = V.kernel.cse.newvar(dtype=x.dtype)
+        exponent = V.kernel.cse.newvar(dtype=torch.int32)
         V.kernel.compute.writeline(
             f"{mantissa}, {exponent} = triton_helpers.frexp({x})"
         )
